lex and yacc: add error checking when yacc is missing
This configuration snippet is copied from app/twm and xgc.
It handles the case when yacc tool is missing but yacc generated code in the
tarball can be used.

Unlike other macros AC_PROG_YACC sets the YACC variable to "yacc"
as a last resort, even if no programs have been found.
Improve error checking by issuing an error message if yacc is not found.
The module cannot build anyway, unless from a published tarball.

If laygram.c is already there, do not abort the configuration
if yacc is missing as it is not needed.

If both generated code and yacc are missing, abort configuration.
Yacc is a required tool to build the package.

The test for YACC_INST is cross-compiling friendly.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>